Move all JPB components from their own folders under a main JobBuilder folder and update all references
