+++ resources/assets/js/components/JobBuilder/Details/JobDetailsMessages.tsx
