+++ resources/assets/js/components/JobBuilder/JobBuilderRoot.tsx
@@ -4,14 +4,14 @@ import ReactDOM from "react-dom";
 import { Routes } from "universal-router";
 import { defineMessages, useIntl } from "react-intl";
 import { useRouter, RouterResult } from "../../helpers/router";
-import JobBuilderIntroPageContainer from "../JobBuilderIntro/JobBuilderIntro";
+import JobBuilderIntroPageContainer from "./Intro/JobBuilderIntro";
 import RootContainer from "../RootContainer";
-import JobDetailsPage from "../JobDetails/JobDetailsPage";
-import JobBuilderWorkEnvPage from "../JobBuilderWorkEnv/JobBuilderWorkEnv";
-import JobBuilderImpactPage from "../JobBuilderImpact/JobBuilderImpactPage";
-import JobTasksPage from "../JobTasks/JobTasksPage";
-import JobSkillsPage from "../JobBuilderSkills/JobBuilderSkillsPage";
-import JobReviewPage from "../JobReview/JobReviewPage";
+import JobDetailsPage from "./Details/JobDetailsPage";
+import JobBuilderWorkEnvPage from "./WorkEnv/JobBuilderWorkEnv";
+import JobBuilderImpactPage from "./Impact/JobBuilderImpactPage";
+import JobTasksPage from "./Tasks/JobTasksPage";
+import JobSkillsPage from "./Skills/JobBuilderSkillsPage";
+import JobReviewPage from "./Review/JobReviewPage";
 import ScrollToTop from "../ScrollToTop";
 import RedirectPage from "./RedirectToLastIncompleteStep";
