+++ resources/assets/js/stories/JobPosterBuilder/Details.stories.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from "@storybook/react";
 import { withIntl } from "storybook-addon-intl";
 import { action } from "@storybook/addon-actions";
 import { boolean } from "@storybook/addon-knobs";
-import { JobDetailsIntl as JobDetails } from "../../components/JobDetails/JobDetails";
+import { JobDetailsIntl as JobDetails } from "../../components/JobBuilder/Details/JobDetails";
 import fakeJob from "../../fakeData/fakeJob";
 
 const stories = storiesOf("Job Poster Builder|Details", module).addDecorator(
