+++ resources/assets/js/stories/JobPosterBuilder/Review.stories.tsx
@@ -6,7 +6,7 @@ import { select } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 import fakeJob, { fakeCriterion, fakeJobTasks } from "../../fakeData/fakeJob";
 import { fakeSkills } from "../../fakeData/fakeSkills";
-import JobReview from "../../components/JobReview/JobReview";
+import JobReview from "../../components/JobBuilder/Review/JobReview";
 import fakeDepartments from "../../fakeData/fakeDepartments";
 import { fakeManager } from "../../fakeData/fakeManager";
