+++ resources/assets/js/stories/JobPosterBuilder/Skills.stories.tsx
@@ -5,10 +5,10 @@ import { withIntl } from "storybook-addon-intl";
 import { select, boolean } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 import { SelectTypeOptionsProp } from "@storybook/addon-knobs/dist/components/types";
-import JobBuilderSkills from "../../components/JobBuilderSkills/JobBuilderSkills";
+import JobBuilderSkills from "../../components/JobBuilder/Skills/JobBuilderSkills";
 import fakeJob, { fakeCriterion, fakeJobTasks } from "../../fakeData/fakeJob";
 import { fakeSkills } from "../../fakeData/fakeSkills";
-import CriteriaForm from "../../components/JobBuilderSkills/CriteriaForm";
+import CriteriaForm from "../../components/JobBuilder/Skills/CriteriaForm";
 import { mapToObject } from "../../helpers/queries";
 import { SkillLevelId, CriteriaTypeId } from "../../models/lookupConstants";
 import { Criteria } from "../../models/types";
